Add new images and home view components
Added new SVG images for filter, notification, and search icon, as well as a PNG image for the profile. Introduced a new `HomeView` and its associated widgets, including `CustomHomeAppbar` and `HomeViewBody`. Updated route generation to include the new `HomeView` and added a search text field widget. Also, modified the sign-in and sign-up view body bloc consumers to handle navigation and error messages appropriately.