Add `isSignedIn` method and update splash navigation logic
Added an `isSignedIn` method to `FirebaseAuthService` to check if a user is signed in. Updated the `excuteNavigation` method in `SplashViewBody` to navigate to the home screen if the user is signed in, otherwise navigate to the sign-in screen. Ensured that the context is checked for being mounted before performing navigation.